Cardslib: fix crash on swipe expanded card and dismiss
Not sure what the idea was behind to collapse the expanded card on dismiss. The user
does not expect to have the card collapsing during it animates away. The swipe
animation takes completly care of it that the animation looks nice.

So we just remove it here.

Problem here was that the card is removed....prepareCollapsedView is called
and get the initial init with the old position. Card gets removed (means positions
are not valid anymore) and then notifiyDataSetChanged is called which finally
call the viewtree observer and animates. Which confuses the whole logic.

Crash is just fixed due that it is not called anymore.

Still the root cause
for the crash needs to be found for the expand/collapse feature itself. Basically
the view recylce confuses the position logic after the animation. A simple notfyDataSetChanged
after the animation solves it but that is a dumb and not performant solution. We may should just
update the animating row and the row below and on top of the animated one

Well patch for this will follow soon.

Change-Id: Idbf45fec954c770623b239df2a5330b89da1d22c