Cardslib: fix parents lose after animation
Currently the parent information get lost after the expand/collapse
animation is done. Results into the problem that on the next animation
in case the view was not fully scrolled at all the animation is broken
or results in a FC when we try to get the view position.

To prevent that call notifyDateSetChanged() after the animation is done
to trigger getView for all visible views and to restore to correct
view hirachy

Well not very happy with this solution. It fixes it for now but there should
be a better and more peformant one. On big lists a notifyDataSetChanded() call
can be resource consuming.

We currently do not use large lists so for now a valid fix.

Change-Id: I243f375a96e09a23be686e7598cb54c7cb437513