Cardslib: update SwipeDismissListViewTouchListener and fix it up
Update to latest improvements from Roman Nurik.

Cleanup a bit

and fix the following issue:

Swipe to the left....then in the same swipe swipe back to the right the view
will jump after going over the initial touchdown point. Reason is that the offset
which was introduced to give a better swipe feeling is taken at this moment into account.

Fix it to set the offset only on swipe start.

As well this commit fixes the sometimes not that responsive areas.

Change-Id: Iafac2db20e5a2044886ae332554d86974773250e